Added a few things since last commit
Fixed the plotting cell, and added extra plots so now the following are "done":
Plotted each of the 4 methods with the analytical solutions shown for comparison
Plotted each method for different step sizes to show increase in accuracy
Plotted each method for different damping coefficients to show increase in accuracy for Cromer and Verlet especially
There's a cell ready to plot the chosen "best" method for varying damping coefficients in terms of the critical damping, b.
Added markdown cells between all of these mentioned code cells with various comments and questions I need to revisit.